# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Typeform Clone API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/public/forms/{slug}": {
      "get": {
        "tags": [
          "Public forms"
        ],
        "summary": "Form Definition",
        "operationId": "form_definition_api_v1_public_forms__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicForm"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/forms/{slug}/responses": {
      "post": {
        "tags": [
          "Public forms"
        ],
        "summary": "Submit Response",
        "operationId": "submit_response_api_v1_public_forms__slug__responses_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResponse"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftDefinition": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionDefinition"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Questions"
          },
          "theme_settings": {
            "$ref": "#/components/schemas/ThemeSettings"
          },
          "ending_settings": {
            "$ref": "#/components/schemas/EndingSettings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DraftDefinition"
      },
      "EndingSettings": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": "Thank you!"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": "Your response has been recorded."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EndingSettings"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptionDefinition": {
        "properties": {
          "option_key": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Option Key"
          },
          "label": {
            "type": "string",
            "maxLength": 500,
            "title": "Label",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "option_key"
        ],
        "title": "OptionDefinition"
      },
...
      "PublicForm": {
        "properties": {
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "definition": {
            "$ref": "#/components/schemas/DraftDefinition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version_id",
          "version_number",
          "definition"
        ],
        "title": "PublicForm"
      },
      "QuestionDefinition": {
        "properties": {
          "question_key": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Question Key"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "title": {
            "type": "string",
            "maxLength": 2000,
            "title": "Title",
            "default": ""
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "rating_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Max"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionDefinition"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_key",
          "type"
        ],
        "title": "QuestionDefinition"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "short_text",
          "long_text",
          "multiple_choice",
          "dropdown",
          "email",
          "number",
          "yes_no",
          "rating"
        ],
        "title": "QuestionType"
      },
...
      "SubmissionReceipt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "version_id",
          "submitted_at"
        ],
        "title": "SubmissionReceipt"
      },
      "SubmitResponse": {
        "properties": {
          "version_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Version Id"
          },
          "submission_key": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Submission Key"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/SubmittedAnswer"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version_id",
          "submission_key",
          "answers"
        ],
        "title": "SubmitResponse"
      },
      "SubmittedAnswer": {
        "properties": {
          "question_key": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Question Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_key",
          "value"
        ],
        "title": "SubmittedAnswer"
      },
      "ThemeSettings": {
        "properties": {
          "font": {
            "type": "string",
            "const": "inter",
            "title": "Font",
            "default": "inter"
          },
          "background": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Background",
            "default": "#ffffff"
          },
          "text": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Text",
            "default": "#2a222b"
          },
          "accent": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Accent",
            "default": "#2563eb"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ThemeSettings"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}